// template/react-starter-kit/src/.generated/api/model/createSupportTicketRequestDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */
import type { SupportTicketPriority } from './supportTicketPriority';

export interface CreateSupportTicketRequestDto {
  /** @maxLength 50 */
  category: string;
  /** @maxLength 255 */
  title: string;
  content: string;
  priority?: SupportTicketPriority;
}